Modification made so that Json can handle large member to allow 10B event running

+++ src/main.cxx
@@ -70,7 +70,9 @@ int main(int argc, char** argv){
    Json::Reader reader;
    reader.parse(ifs, obj);
  
-   int nEvents = obj["n_events"].asInt();
+//   int nEvents = obj["n_events"].asInt();
+//   int nEvents = obj["n_events"].asUInt64();
+   signed long long int nEvents = obj["n_events"].asInt64();
    cout << "Generating "<< nEvents << " events."<<endl;
 
    TString file_name = obj["file_name"].asString();
